move existing VM helpers into a dedicated file
