+++ test/util/framework/hcp_helper.go
@@ -27,7 +27,6 @@ import (
 	"fmt"
 	"net/http"
 	"regexp"
-	"strings"
 	"time"
 
 	"github.com/davecgh/go-spew/spew"
@@ -51,7 +50,6 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
-	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/compute/armcompute/v5"
 
 	"github.com/Azure/ARO-HCP/internal/api"
 	hcpsdk20240610preview "github.com/Azure/ARO-HCP/test/sdk/resourcemanager/redhatopenshifthcp/armredhatopenshifthcp"
@@ -1032,92 +1030,6 @@ func BuildNodePoolFromParams(
 	return nodePool
 }
 
-// Helper to run command on VM
-func RunVMCommand(ctx context.Context, tc interface {
-	SubscriptionID(ctx context.Context) (string, error)
-	AzureCredential() (azcore.TokenCredential, error)
-}, resourceGroup, vmName, command string, pollTimeout time.Duration) (string, error) {
-	subscriptionID, err := tc.SubscriptionID(ctx)
-	if err != nil {
-		return "", err
-	}
-
-	azCreds, err := tc.AzureCredential()
-	if err != nil {
-		return "", err
-	}
-
-	computeClient, err := armcompute.NewVirtualMachinesClient(subscriptionID, azCreds, nil)
-	if err != nil {
-		return "", err
-	}
-
-	runCommandInput := armcompute.RunCommandInput{
-		CommandID: to.Ptr("RunShellScript"),
-		Script: []*string{
-			to.Ptr(command),
-		},
-	}
-
-	poller, err := computeClient.BeginRunCommand(ctx, resourceGroup, vmName, runCommandInput, nil)
-	if err != nil {
-		return "", err
-	}
-
-	// Create a timeout context to avoid waiting too long on VM command failures
-	// VM commands should complete quickly (within a few minutes at most)
-	pollCtx, cancel := context.WithTimeout(ctx, pollTimeout)
-	defer cancel()
-
-	result, err := poller.PollUntilDone(pollCtx, nil)
-	if err != nil {
-		return "", err
-	}
-
-	if len(result.Value) > 0 && result.Value[0].Message != nil {
-		// Azure Run Command returns output in format:
-		// "Enable succeeded: \n[stdout]\n<actual output>\n[stderr]\n<errors>"
-		// We need to extract stdout and stderr content
-		message := *result.Value[0].Message
-
-		// Find the stdout section
-		stdoutStart := strings.Index(message, "[stdout]\n")
-		if stdoutStart == -1 {
-			// If no stdout marker, return the whole message
-			return message, nil
-		}
-
-		// Skip past the "[stdout]\n" marker
-		stdoutStart += len("[stdout]\n")
-
-		// Find where stderr starts (if present)
-		stderrStart := strings.Index(message[stdoutStart:], "\n[stderr]")
-
-		var output string
-		if stderrStart == -1 {
-			// No stderr marker, take everything after stdout
-			output = message[stdoutStart:]
-		} else {
-			// Take only the stdout section
-			output = message[stdoutStart : stdoutStart+stderrStart]
-
-			// Extract and inspect stderr
-			stderrAbsoluteStart := stdoutStart + stderrStart + len("\n[stderr]\n")
-			if stderrAbsoluteStart < len(message) {
-				stderr := strings.TrimSpace(message[stderrAbsoluteStart:])
-				if stderr != "" {
-					// Return an error if stderr is not empty
-					return "", fmt.Errorf("%s", stderr)
-				}
-			}
-		}
-
-		return strings.TrimSpace(output), nil
-	}
-
-	return "", nil
-}
-
 // Helper to generate SSH key pair
 func GenerateSSHKeyPair() (publicKey string, privateKey string, err error) {
 	// Generate RSA key pair
@@ -1467,48 +1379,6 @@ func GetNodePool20251223(
 		return nil, err
 	}
 	return &resp.NodePool, nil
-}
-
-// GetVirtualMachinesInResourceGroup lists all VMs in the given resource group
-// with a polling loop to handle ARM replication delays.
-func GetVirtualMachinesInResourceGroup(
-	ctx context.Context,
-	computeClientFactory *armcompute.ClientFactory,
-	resourceGroupName string,
-	expectedMinimumCount int,
-	timeout time.Duration,
-) ([]*armcompute.VirtualMachine, error) {
-	ctx, cancel := context.WithTimeoutCause(ctx, timeout,
-		fmt.Errorf("timed out waiting for %d VMs in resource group %q", expectedMinimumCount, resourceGroupName))
-	defer cancel()
-
-	vmClient := computeClientFactory.NewVirtualMachinesClient()
-	const pollInterval = 10 * time.Second
-
-	for {
-		var vms []*armcompute.VirtualMachine
-		pager := vmClient.NewListPager(resourceGroupName, nil)
-		for pager.More() {
-			page, err := pager.NextPage(ctx)
-			if err != nil {
-				if errors.Is(err, context.DeadlineExceeded) {
-					return nil, fmt.Errorf("caused by: %w, error: %w", context.Cause(ctx), err)
-				}
-				return nil, fmt.Errorf("failed to list VMs in resource group %q: %w", resourceGroupName, err)
-			}
-			vms = append(vms, page.Value...)
-		}
-
-		if len(vms) >= expectedMinimumCount {
-			return vms, nil
-		}
-
-		select {
-		case <-ctx.Done():
-			return nil, fmt.Errorf("caused by: %w, error: %w", context.Cause(ctx), ctx.Err())
-		case <-time.After(pollInterval):
-		}
-	}
 }
 
 // BuildHCPClusterFromParamsV20251223 builds a v20251223preview HCP cluster from ClusterParamsV20251223
